Avance proyecto
Se modifico las estructuras de las diferentes carpetas. También se  incorporaron estilos

+++ src/public/js/script.js
@@ -11,37 +11,53 @@ let idForm = "";
 const inputTitle = document.getElementById("inputTitle");
 const inputDescription = document.getElementById("inputDescription");
 const inputPoster = document.getElementById("inputPoster");
+var post = {createdAT: new Date(),};
+
 
 btnCrear.addEventListener("click", () => {
   option = "new";
-  btnSave.textContent = "new";
+  btnSave.textContent = "Publicar";
   inputTitle.value = "";
   inputDescription.value = "";
   inputPoster.value = "";
   myModal.show();
 });
 
+document.addEventListener("click", (event) => {
+  if (event.target.matches("#btn-delete")) {
+    const article = event.target.closest(".col-6");
+    const idArticle = article.dataset.id;
 
-document.addEventListener('click', (event) => {
-    if (event.target.matches('#btn-delete')) {
-        const article = event.target.closest('.col-4')
-        const idArticle = article.dataset.id
-
+    Swal.fire({
+      title: "Esta seguro de querer borrar la publicación?",
+      text: "Se perdera todo el contenido!",
+      icon: "Alerta",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Si! Borrar!",
+    }).then((result) => {
+      if (result.isConfirmed) {
         fetch(`http://localhost:3000/api/tasks/${idArticle}`, {
-            method: 'DELETE'
-        }).then(res => {
-            if (res.ok) {
-                article.remove()
-            }
-        }).catch(err => {
-            console.error(err)
+          method: "DELETE",
         })
-    }
-})
+          .then((res) => {
+            if (res.ok) {
+              article.remove();
+            }
+          })
+          .catch((err) => {
+            console.error(err);
+          });
+        Swal.fire("Borrado!", "Su publicación fue eliminada.", "success");
+      }
+    });
+  }
+});
 
 document.addEventListener("click", (event) => {
   if (event.target.matches("#btn-edit")) {
-    const article = event.target.closest(".col-4");
+    const article = event.target.closest(".col-6");
 
     const idArticle = article.dataset.id;
     const urlPosterEdit = article.children[0].children[0].src;
@@ -54,15 +70,14 @@ document.addEventListener("click", (event) => {
     inputDescription.value = descriptionEdit;
     inputPoster.value = urlPosterEdit;
     option = "edit";
-    btnSave.textContent = "Edit";
+    btnSave.textContent = "Editar";
     myModal.show();
   }
 });
 
 form.addEventListener("submit", (event) => {
   event.preventDefault();
-  // console.log("submit");
-
+  
   if (option === "new") {
     const newTask = {
       title: inputTitle.value,
@@ -79,7 +94,7 @@ form.addEventListener("submit", (event) => {
     }).then(res => {
       console.log(res)
         if (res.ok) {
-          alert("Task created successfully");
+          alert("Publicación creada con exito");
           myModal.hide();
           location.reload();
         }
@@ -92,8 +107,8 @@ form.addEventListener("submit", (event) => {
   if (option === "edit") {
     const newTask = {
       title: inputTitle.value,
-      description: inputDescription.value,
       poster: inputPoster.value,
+      description: inputDescription.value,
     };
 
     fetch(`http://localhost:3000/api/tasks/${idForm}`, {
@@ -104,7 +119,7 @@ form.addEventListener("submit", (event) => {
       body: JSON.stringify(newTask)
     }).then(res => {
       if(res.ok){
-        alert('Task edited successfully')
+        alert('Publicación editada')
         myModal.hide();
         location.reload();
       }
@@ -112,3 +127,32 @@ form.addEventListener("submit", (event) => {
   }
 });
 
+fetch("http://localhost:3000/api/tasks")
+  .then((res) => res.json())
+  .then((data) => {
+     data.forEach((task) => {
+      html += `
+            <article class="col-6 d-flex justify-content-center mb-3" data-id="${task.id}">
+                <div class="card" style="width: 26rem;">
+                    <img src="${task.poster}"
+                        class="card-img-top" alt="nuevo titulo">
+                    <div class="card-body">
+                        <h5 class="card-title">${task.title}</h5>
+                        <p class="card-text">${task.description}</p>
+                        <div>
+                            <button class="btn btn-secondary" id="btn-edit">Editar</button>
+                            <button type="" class="btn btn-danger" id="btn-delete">Borrar</button>
+                        </div>
+                        <div>
+                            <p class="card-fecha d-flex justify-content-end" id="fecha">
+                                ${new Date(post.createdAT).toLocaleDateString()}
+                             </p>
+                        </div>
+                    </div>
+                </div>
+            </article>
+            `;
+
+      contenedor.innerHTML = html;
+    });
+  });
